orginize configs
