+++ src/config/board.ts
@@ -1,0 +1,5 @@
+export const CIRCLE_RADIUS = 5;
+export const CIRCLE_DIAMETER = 2 * CIRCLE_RADIUS;
+export const CANVAS_WIDTH = 600;
+export const CANVAS_HEIGHT = 600;
+export const PADDING = 30;
